+++ app/src/main/java/com/example/cooklette/database/entity/RecipeIngredient.kt
@@ -3,17 +3,16 @@ package com.example.cooklette.database.entity
 import androidx.annotation.NonNull
 import androidx.room.ColumnInfo
 import androidx.room.Entity
-import androidx.room.ForeignKey
 
 @Entity(primaryKeys = ["id_recipe", "id_ingredient"])
 data class RecipeIngredient(
     @NonNull
-    val id_recipe: Int,
+    val id_recipe: Long,
     @NonNull
     @ColumnInfo(index = true)
-    val id_ingredient: Int,
+    val id_ingredient: Long,
     @NonNull
-    val id_unit: Int,
+    val id_unit: Long,
     @NonNull
-    val quantity: Int
+    val quantity: Long
 )
